% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-alluvium.r
\name{data_to_alluvium}
\alias{data_to_alluvium}
\title{Create alluvium curve points from data}
\usage{
data_to_alluvium(
  data,
  curve_type = "spline",
  knot.prop = TRUE,
  curve_range = 6,
  segments = 3
)
}
\description{
Placeholder. Documentation goes here.
}
