% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-lode.r
\docType{data}
\name{geom-lode}
\alias{geom-lode}
\alias{geom_lode}
\alias{GeomLode}
\title{Lode rectangles}
\format{An object of class \code{GeomLode} (inherits from \code{Geom}, \code{ggproto}) of length 7.}
\arguments{
\item{stat}{The statistical transformation to use on the data;
override the default.}

\item{width}{Numeric; the width of each stratum, as a proportion of the
distance between axes. Defaults to 1/3.}

\item{axis_width}{Deprecated; alias for \code{width}.}
}
\description{
\code{geom_alluvium} receives a dataset of the horizontal (\code{x}) and 
vertical (\code{y}, \code{ymin}, \code{ymax}) positions of the \strong{lodes}
of an alluvial diagram, the intersections of the alluvia with the strata.
It plots rectangles for these lodes of a provided \code{width}.
}
\section{Aesthetics}{

\code{geom_lode} understands the following aesthetics
  (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{y}}
  \item \strong{\code{ymin}}
  \item \strong{\code{ymax}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{linetype}
  \item \code{size}
  \item \code{group}
}
Currently, \code{group} is ignored.
}

\examples{
# one axis
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis = Class)) +
  geom_lode(aes(fill = Class, alpha = Survived)) +
  scale_x_continuous(breaks = 1, labels = c("Class")) +
  scale_alpha_manual(values = c(.25, .75))

# alluvia and lodes
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age,
           fill = Survived)) +
  geom_alluvium() + geom_lode()

# lodes as strata
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age,
           fill = Survived)) +
  geom_alluvium() +
  geom_stratum(stat = "alluvium")
}
\seealso{
\code{\link[ggplot2]{layer}} for additional arguments,
  \code{\link{geom_alluvium}} for alluvial flows,
  \code{\link{stat_stratum}} and \code{\link{geom_stratum}} for 
  intra-axis stratum rectangles, and
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
