% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-axes.r
\name{get_axes}
\alias{get_axes}
\title{Identify axis columns}
\usage{
get_axes(data)
}
\arguments{
\item{data}{data.frame}
}
\description{
Identify variables in a dataset that fit the pattern of axis aesthetic names,
and return their column indices in the numerical order of the axis numbers.
}

