% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r, R/ggalluvial.r
\name{ggalluvial-deprecated}
\alias{ggalluvial-deprecated}
\alias{ggalluvial}
\alias{ggalluvial_lodes}
\alias{ggalluvial_alluvia}
\alias{ggalluvial_formula}
\title{Deprecated functions}
\usage{
ggalluvial(...)

ggalluvial_lodes(..., incl.strata = TRUE, incl.labels = FALSE)

ggalluvial_alluvia(..., incl.strata = TRUE, incl.labels = FALSE)

ggalluvial_formula(formula, data = NULL, weight, incl.strata = TRUE,
  incl.labels = TRUE, ...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot} and inherited by 
\code{geom_alluvium} and \code{geom_stratum}.}

\item{incl.strata}{Logical; whether to plot strata over alluvia.}

\item{incl.labels}{Logical; whether to plot labels over strata.}

\item{formula}{A formula to specify the axes and alluvial divisions.}

\item{data}{A data frame or frequency table.}

\item{weight}{A weight variable, from \code{data} or of compatible length 
with the elements of \code{formula}.}
}
\description{
These functions are deprecated in the current version and will be removed for
the eventual CRAN release.
}
\seealso{
\code{\link{stat_stratum}}, \code{\link{stat_flow}},
  \code{\link{stat_alluvium}}, \code{\link{geom_stratum}},
  \code{\link{geom_lode}}, \code{\link{geom_flow}}, and
  \code{\link{geom_alluvium}}.
}
\keyword{internal}
