% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-alluvium.r, R/geom-flow.r, R/geom-lode.r,
%   R/geom-stratum.r, R/ggproto.r, R/stat-alluvium.r, R/stat-flow.r,
%   R/stat-stratum.r
\name{GeomAlluvium}
\alias{GeomAlluvium}
\alias{GeomFlow}
\alias{GeomLode}
\alias{GeomStratum}
\alias{ggalluvial-ggproto}
\alias{StatAlluvium}
\alias{StatFlow}
\alias{StatStratum}
\title{Base ggproto classes for ggalluvial}
\description{
Base ggproto classes for ggalluvial
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}}
}
\keyword{internal}
