% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalluvial-package.r
\docType{package}
\name{ggalluvial-package}
\alias{ggalluvial}
\alias{ggalluvial-package}
\title{ggalluvial: Alluvial Plots in 'ggplot2'}
\description{
Alluvial plots use variable-width ribbons and stacked bar plots to represent multi-dimensional or repeated-measures data with categorical or ordinal variables; see Riehmann, Hanfler, and Froehlich (2005) \doi{10.1109/INFVIS.2005.1532152} and Rosvall and Bergstrom (2010) \doi{10.1371/journal.pone.0008694}. Alluvial plots are statistical graphics in the sense of Wilkinson (2006) \doi{10.1007/0-387-28695-0}; they share elements with Sankey diagrams and parallel sets plots but are uniquely determined from the data and a small set of parameters. This package extends Wickham's (2010) \doi{10.1198/jcgs.2009.07098} layered grammar of graphics to generate alluvial plots from tidy data.
}
\section{Acknowledgments}{


Many users identified problems and suggested improvements via email and the
GitHub issue tracker.

Development benefitted from the use of equipment and the support of
colleagues at \href{https://health.uconn.edu/}{UConn Health} and at \href{https://ufhealth.org/}{UF Health}.
}

\seealso{
Useful links:
\itemize{
  \item \url{http://corybrunson.github.io/ggalluvial/}
  \item Report bugs at \url{https://github.com/corybrunson/ggalluvial/issues}
}

}
\author{
\strong{Maintainer}: Jason Cory Brunson \email{cornelioid@gmail.com}

Authors:
\itemize{
  \item Jason Cory Brunson \email{cornelioid@gmail.com}
  \item Quentin D. Read
}

}
\keyword{internal}
