% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalluvial.r
\name{ggalluvial}
\alias{ggalluvial}
\title{Quick alluvial diagram}
\usage{
ggalluvial(...)
}
\arguments{
\item{...}{arguments passed to \code{ggplot} and inherited by
\code{geom_alluvium} and \code{geom_stratum}.}
}
\description{
Produces an alluvial diagram with axis strata and labels.
}
\seealso{
\code{\link{geom_alluvium}} and \code{\link{geom_stratum}}
}

