% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-alluvial.r
\name{is_alluvial}
\alias{is_alluvial}
\title{Check a data frame for alluvial structure}
\usage{
is_alluvial(data, ..., logical = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{logical}{Whether to return a logical value (TRUE, the default) or a
character string indicating the type of alluvial structure ("none",
"lodes", or "alluvia")}
}
\description{
Alluvial diagrams consist of multiple horizontally-distributed columns (axes)
representing factor variables, vertical divisions (strata) of these axes
representing these variables' values; and splines (alluvial flows) connecting
vertical subdivisions (lodes) within strata of adjacent axes representing
subsets or amounts of observations that take the corresponding values of the
corresponding variables. This function checks a data frame for either of two
types of alluvial structure:
\itemize{
  \item One row per **lode**, wherein each row encodes a subset or amount of
        observations having a specific profile of axis values, a `key` field
        encodes the axis, a `value` field encodes the value within each axis,
        and a `id` column identifies multiple lodes corresponding to the
        same subset or amount of observations.
  \item One row per **alluvium**, wherein each row encodes a subset or amount
        of observations having a specific profile of axis values and a set
        `axes` of fields encodes its values at each axis variable.
}
}
