% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lode-guidance-functions.r
\name{lode-guidance-functions}
\alias{lode-guidance-functions}
\alias{lode_leftright}
\alias{lode_leftward}
\alias{lode_rightleft}
\alias{lode_rightward}
\alias{lode_zigzag}
\title{Lode guidance functions}
\usage{
lode_zigzag(n, i)

lode_rightward(n, i)

lode_leftward(n, i)

lode_rightleft(n, i)

lode_leftright(n, i)
}
\arguments{
\item{n}{Numeric, a positive integer}

\item{i}{Numeric, a positive integer at most \code{n}}
}
\description{
Order the numbers 1 through \code{n}, starting at some index and zigzagging
outward.

Order the numbers 1 through \code{n}, starting at some index \code{i} and
putting the remaining indices in increasing order.

Order the numbers 1 through \code{n}, starting at some index \code{i} and
putting the remaining indices in decreasing order.

Order the numbers 1 through \code{n}, starting at some index \code{i},
proceeding rightward to \code{n}, then proceeding leftward to \code{1}.

Order the numbers 1 through \code{n}, starting at some index \code{i},
proceeding leftward to \code{1}, then proceeding rightward to \code{n}.
}

