% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lode_leftward.r
\name{lode_leftward}
\alias{lode_leftward}
\title{Leftward outward order}
\usage{
lode_leftward(n, i)
}
\arguments{
\item{n}{Numeric, a positive integer}

\item{i}{Numeric, a positive integer at most \code{n}}
}
\description{
Order the numbers 1 through \code{n}, starting at some index \code{i} and
putting the remaining indices in decreasing order.
}

