% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lode_rightward.r
\name{lode_rightward}
\alias{lode_rightward}
\title{Rightward outward order}
\usage{
lode_rightward(n, i)
}
\arguments{
\item{n}{Numeric, a positive integer}

\item{i}{Numeric, a positive integer at most \code{n}}
}
\description{
Order the numbers 1 through \code{n}, starting at some index \code{i} and
putting the remaining indices in increasing order.
}

