% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lode_zigzag.r
\name{lode_zigzag}
\alias{lode_zigzag}
\title{Zigzag outward order}
\usage{
lode_zigzag(n, i)
}
\arguments{
\item{n}{Numeric, a positive integer}

\item{i}{Numeric, a positive integer at most \code{n}}
}
\description{
Order the numbers 1 through \code{n}, starting at some index and zigzagging
outward.
}

