% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{nsa}
\alias{nsa}
\title{Influenza vaccination rates}
\format{An alluvial data frame.}
\source{
RAND American Life Panel \url{https://alpdata.rand.org/}
}
\description{
This is a sample from the RAND ALP surveys on influenza vaccination.
}
