% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-strata.r
\name{optimize_strata}
\alias{optimize_strata}
\title{Optimize strata}
\usage{
optimize_strata(data, key, value, id, weight = NULL, objective = "count",
  method = "exhaustive", niter = 10)
}
\arguments{
\item{data}{Data frame in lode form (see \code{\link{to_lodes}}).}

\item{key, value, id}{Numeric or character; the fields of \code{data}
corresponding to the axis (variable), stratum (value), and alluvium
(identifying) variables.}

\item{weight}{Optional numeric or character; the fields of \code{data}
corresponding to alluvium or lode weights (heights when plotted). The
default value \code{NULL} assigns every alluvium equal weight.}

\item{objective}{Character, the objective function to minimize; matched to
\code{"count"}, \code{"weight"}, or \code{"area"}.}

\item{method}{Character; whether to exhaust all permutations of all axes
(\code{"exhaustive"}) versus to perform heuristic optimization from random
initial permutations (\code{"heuristic"}).}

\item{niter}{Positive integer; if \code{method == "heuristic"}, the number of
iterations to perform from randomly sampled seed permutation sets.}
}
\description{
This function attempts to order the strata at each axis so as to maximize the
readability of the diagram, by minimizing an objective function related to
the number and sizes of the overlaps between flows when they cross. The
planned objective functions include the following:
\itemize{
  \item \code{"count"}: the number of crossings
  \item \code{"weight"}: the total weight of the crossings
  (the weight of each crossing is the product of the weights of the flows)
  \item \code{"area"}: the total area of the crossings
  (weights have less area when the flows intersect at greater angles)
}
}
