% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-stratum.r
\name{rect_to_poly}
\alias{rect_to_poly}
\title{Convert rectangle to polygon
(lifted from \code{\link[ggplot2]{geom_rect}})}
\usage{
rect_to_poly(xmin, xmax, ymin, ymax)
}
\description{
Convert rectangle to polygon
(lifted from \code{\link[ggplot2]{geom_rect}})
}
