% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinters.r
\name{splinters}
\alias{splinters}
\title{Check that one subset decomposition splinters another}
\usage{
splinters(x, y)
}
\arguments{
\item{x}{Vector of any class; encodes the divisor.}

\item{y}{Vector of any class; encodes the dividend.}
}
\description{
Check that the distinct values of one vector splinter those of another
vector, in the sense that the binning defined by the former is finer (or
equal to) that of the latter, i.e. each bin in the former is a subset of one
in the latter.
}
