% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-alluvium.r
\docType{data}
\name{stat-alluvium}
\alias{stat-alluvium}
\alias{stat_alluvium}
\alias{StatAlluvium}
\title{Alluvial positions}
\format{An object of class \code{StatAlluvium} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\arguments{
\item{geom}{The geometric object to use display the data;
override the default.}

\item{decreasing}{Logical; whether to arrange the strata at each axis
in the order of the variable values (NA, the default),
with the largest on top (FALSE), or
with the largest on bottom (TRUE).}

\item{lode.guidance}{The function to prioritize the axis variables for 
ordering the lodes within each stratum. Defaults to "zigzag", other options
include "rightleft", "leftright", "rightward", and "leftward" (see 
\code{\link{lode-guidance-functions}}).}

\item{aes.bind}{Whether to prioritize aesthetics before axes (other than the
index axis) when ordering the lodes within each stratum. Defaults to FALSE.}

\item{lode.ordering}{A list (of length the number of axes) of integer vectors
(each of length the number of rows of \code{data}) or NULL entries 
(indicating no imposed ordering), or else a numeric matrix of corresponding
dimensions, giving the preferred ordering of alluvia at each axis. This 
will be used to order the lodes within each stratum by sorting the lodes 
first by stratum and then by the provided vectors.}
}
\description{
Given a dataset with alluvial structure, \code{stat_alluvium} calculates the 
centroids (\code{x} and \code{y}) of the \strong{lodes}, the intersections of
the alluvia with the strata, together with their weights (heights; 
\code{ymin} and \code{ymax}). It leverages the \code{group} aesthetic for 
plotting purposes (for now).
}
\section{Aesthetics}{
 \code{stat_alluvium} understands the following
  aesthetics (required aesthetics are in bold):
\itemize{
  \item \code{x}
  \item \code{stratum}
  \item \code{alluvium}
  \item \code{axis[0-9]*} (\code{axis1}, \code{axis2}, etc.)
  \item \code{weight}
  \item \code{group}
}
Currently, \code{group} is ignored.
Use \code{x}, \code{stratum}, and \code{alluvium} for data in lode form and 
\code{axis[0-9]*} for data in alluvium form (see \code{\link{is_alluvial}});
arguments to parameters inconsistent with the data format will be ignored.
}

\examples{
# illustrate positioning
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age,
           color = Survived)) +
  geom_errorbar(stat = "stratum") +
  geom_line(stat = "alluvium") +
  geom_pointrange(stat = "alluvium") +
  geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:3,
                     labels = c("Class", "Sex", "Age"))
}
\seealso{
\code{\link[ggplot2]{layer}} for additional arguments,
  \code{\link{geom_alluvium}} for the corresponding geom,
  \code{\link{stat_stratum}} and \code{\link{geom_stratum}} for
  intra-axis boxes, and
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
