% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-flow.r
\docType{data}
\name{stat-flow}
\alias{stat-flow}
\alias{stat_flow}
\alias{StatFlow}
\title{Flow positions}
\format{An object of class \code{StatFlow} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\arguments{
\item{geom}{The geometric object to use display the data;
override the default.}

\item{decreasing}{Logical; whether to arrange the strata at each axis
in the order of the variable values (NA, the default),
with the largest on top (FALSE), or
with the largest on bottom (TRUE).}

\item{aggregate.wts}{Logical; whether to aggregate weights across otherwise
equivalent rows before computing lode and flow positions. Defaults to TRUE.}

\item{aes.bind}{Whether to prioritize aesthetics before axes (other than the
index axis) when ordering the lodes within each stratum. Defaults to FALSE.}
}
\description{
Given a dataset with alluvial structure, \code{stat_flow} calculates the
centroids (\code{x} and \code{y}) and weights (heights; \code{ymin} and
\code{ymax}) of alluvial flows between each pair of adjacent axes.
}
\section{Aesthetics}{
 \code{stat_flow} understands the following aesthetics
  (required aesthetics are in bold):
\itemize{
  \item \code{x}
  \item \code{stratum}
  \item \code{alluvium}
  \item \code{axis[0-9]*} (\code{axis1}, \code{axis2}, etc.)
  \item \code{weight}
  \item \code{group}
}
Currently, \code{group} is ignored.
Use \code{x}, \code{stratum}, and \code{alluvium} for data in lode form and 
\code{axis[0-9]*} for data in alluvium form (see \code{\link{is_alluvial}});
arguments to parameters inconsistent with the data format will be ignored.
}

\examples{
# illustrate positioning
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age,
           color = Survived)) +
  geom_errorbar(stat = "stratum") +
  geom_line(stat = "flow") +
  geom_pointrange(stat = "flow") +
  geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:3,
                     labels = c("Class", "Sex", "Age"))

data(vaccinations)
# rightward alluvial aesthetics for vaccine survey data
ggplot(vaccinations,
       aes(x = survey, stratum = response, alluvium = subject,
           weight = freq, fill = response)) +
  geom_flow(stat = "alluvium", lode.guidance = "rightward") +
  geom_stratum(alpha = .5) +
  geom_text(stat = "stratum")
# memoryless flows for vaccine survey data
ggplot(vaccinations,
       aes(x = survey, stratum = response, alluvium = subject,
           weight = freq, fill = response)) +
  geom_flow() +
  geom_stratum(alpha = .5) +
  geom_text(stat = "stratum")

# use of lode controls
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_flow(aes(fill = Survived), aes.bind = TRUE) +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))

# both partition aesthetics and interpolation aesthetics
data(vaccinations)
vaccinations$subgroup <- LETTERS[1:2][rbinom(
  n = length(unique(vaccinations$subject)), size = 1, prob = .5
) + 1][vaccinations$subject]
ggplot(vaccinations,
       aes(x = survey, stratum = response, alluvium = subject,
           weight = freq, fill = response)) +
  geom_flow(aes(alpha = subgroup)) +
  scale_alpha_discrete(range = c(1/3, 2/3)) +
  geom_stratum(alpha = .5) +
  geom_text(stat = "stratum")
}
\seealso{
\code{\link[ggplot2]{layer}} for additional arguments,
  \code{\link{geom_flow}} for the corresponding geom,
  \code{\link{stat_stratum}} and \code{\link{geom_stratum}} for
  intra-axis boxes, and
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
