% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-stratum.r
\docType{data}
\name{stat-stratum}
\alias{stat-stratum}
\alias{stat_stratum}
\alias{StatStratum}
\title{Variable axes with strata of values}
\format{An object of class \code{StatStratum} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\arguments{
\item{width}{The width of each stratum, as a proportion of the separation
between their centers. Defaults to 1/3.}

\item{axis_width}{Deprecated; alias for \code{width}.}
}
\description{
Given a dataset with alluvial structure, \code{stat_stratum} calculates the
centroids of the strata for each axis, together with their weights (heights)
and widths.
}
\section{Aesthetics}{

\code{stat_stratum} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
  \item \code{x}
  \item \code{stratum}
  \item \code{alluvium}
  \item \code{axis[0-9]*} (\code{axis1}, \code{axis2}, etc.)
  \item \code{weight}
  \item \code{group}
}
Currently, \code{group} is ignored.
Use \code{x}, \code{stratum}, and \code{alluvium} for data in lode form and 
\code{axis[0-9]*} for data in alluvium form (see \code{\link{is_alluvial}});
arguments to parameters inconsistent with the data form will be ignored.
}

\seealso{
\code{\link{geom-stratum}} for the corresponding geom.
}
\keyword{datasets}
