% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-stratum.r
\docType{data}
\name{StatStratum}
\alias{StatStratum}
\title{Variable axes and strata}
\format{An object of class \code{StatStratum} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\usage{
StatStratum
}
\arguments{
\item{axis_width}{The width of each variable axis, as a proportion of the
separation between axes.}
}
\description{
\code{stat_stratum} calculates the centers of the levels at each axis.
\code{geom_stratum} stacks a box for each level of a variable at its axis.
}
\seealso{
\code{\link{geom_alluvium}} for inter-axis flows.
}
\keyword{datasets}

