% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-alluvia.r
\name{to_alluvia}
\alias{to_alluvia}
\title{Expand an axis indicator and an axis value into multiple axis variables}
\usage{
to_alluvia(data, key, value, id)
}
\arguments{
\item{data}{Data frame.}

\item{key, value, id}{Numeric or character; the fields corresponding to the
axis (variable), stratum (value), and alluvium (identifying) variables.}
}
\description{
Take a data frame with axis and axis value variables to be used in an 
alluvial diagram, and reshape the data frame so that the axes constitute 
separate variables whose values are given by the value variable. This 
function invokes \code{\link[tidyr]{spread_}}.
}
