% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-lodes.r
\name{to_lodes}
\alias{to_lodes}
\title{Combine multiple axis variables into an axis indicator and an axis value}
\usage{
to_lodes(data, key = "x", value = "stratum", id = "alluvium", axes)
}
\arguments{
\item{data}{Data frame.}

\item{key, value, id}{Character; names given to the axis (variable), stratum
(value), and alluvium (identifying) variables.}

\item{axes}{Numeric or character vector; which variables to use as axes.}
}
\description{
Take a data frame with several designated variables to be used as axes in an 
alluvial diagram, and reshape the data frame so that the axis variable names 
constitute a new factor variable and their values comprise another. Other 
variables' values will be repeated, and a row-grouping variable can be 
introduced. This function invokes \code{\link[tidyr]{gather}}.
}
