% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{vaccinations}
\alias{vaccinations}
\title{Influenza vaccination rates}
\format{
An alluvial data frame in lodes form.
}
\source{
RAND American Life Panel \url{https://alpdata.rand.org/}
}
\description{
This data set is aggregated from a longitudinal sample from three RAND
American Life Panel surveys that included questions on influenza vaccination.
The data were kindly contributed by Raffaele Vardavas.
}
