% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigzag.r
\name{zigzag}
\alias{zigzag}
\title{Zigzag outward order}
\usage{
zigzag(n, i)
}
\arguments{
\item{n}{numeric, a positive integer}

\item{i}{numeric, a positive integer at most \code{n}}
}
\description{
Order the numbers 1 through \code{n}, starting at some index and zigzagging
outward.
}

