% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial-data.r
\name{alluvial-data}
\alias{alluvial-data}
\alias{is_alluvial}
\alias{is_alluvial_lodes}
\alias{is_alluvial_alluvia}
\alias{to_lodes}
\alias{to_alluvia}
\title{Check for alluvial structure and convert between alluvial formats}
\usage{
is_alluvial(data, ..., logical = TRUE, silent = FALSE)

is_alluvial_lodes(data, key, value, id, weight = NULL, logical = TRUE,
  silent = FALSE)

is_alluvial_alluvia(data, axes = NULL, weight = NULL, logical = TRUE,
  silent = FALSE)

to_lodes(data, key = "x", value = "stratum", id = "alluvium", axes,
  keep = FALSE, relevel.strata = NULL)

to_alluvia(data, key, value, id)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Additional parameters used to determine method and passed
thereto. All or none of \code{key}, \code{value}, and \code{id}, or else
optionally \code{axes}, and (in either case) optionally \code{weight}.}

\item{logical}{Whether to return a logical value or a character string
indicating the type of alluvial structure ("none", "lodes", or "alluvia")}

\item{silent}{Whether to print warning messages.}

\item{key, value, id}{Numeric or character; the fields of \code{data}
corresponding to the axis (variable), stratum (value), and alluvium
(identifying) variables.}

\item{weight}{Optional numeric or character; the fields of \code{data}
corresponding to alluvium or lode weights (heights when plotted).}

\item{axes}{Numeric or character vector; the field(s) of \code{data}
corresponding to the axi(e)s (variable(s)).}

\item{keep}{Either a numeric or character vector; which variables among those
passed to \code{axes} to merge into the reshapen data by \code{id}, or
logical, indicating whether to merge all (\code{TRUE}) or none
(\code{FALSE}) of these variables.}

\item{relevel.strata}{Optional logical value or character vector to control 
the order of the factor levels of the combined stratum variable. If 
logical, \code{FALSE} keeps the levels of the combined strata in order of 
appearance while \code{TRUE} \code{sort()}s the levels. If character,
interpreted as an ordered vector of factor levels, after which any
additional levels will be appended.}
}
\description{
Alluvial diagrams consist of multiple horizontally-distributed columns (axes)
representing factor variables, vertical divisions (strata) of these axes
representing these variables' values; and splines (alluvial flows) connecting
vertical subdivisions (lodes) within strata of adjacent axes representing
subsets or amounts of observations that take the corresponding values of the
corresponding variables. This function checks a data frame for either of two
types of alluvial structure:
\itemize{
  \item One row per \strong{lode}, wherein each row encodes a subset or
        amount of observations having a specific profile of axis values, a
        \code{key} field encodes the axis, a \code{value} field encodes the
        value within each axis, and a \code{id} column identifies multiple
        lodes corresponding to the same subset or amount of observations.
  \item One row per \strong{alluvium}, wherein each row encodes a subset or
        amount of observations having a specific profile of axis values and a
        set \code{axes} of fields encodes its values at each axis variable.
}
If no arguments are assigned to any of these parameters, then
\code{is_alluvial} will default to \code{is_alluvial_alluvia} and assume that
all fields in \code{data} (other than \code{weight}, if given) are to be
treated as axes.
}
\details{
\code{to_lodes} takes a data frame with several designated variables to be 
used as axes in an alluvial diagram, and reshapes the data frame so that the 
axis variable names constitute a new factor variable and their values 
comprise another. Other variables' values will be repeated, and a 
row-grouping variable can be introduced. This function invokes 
\code{\link[tidyr]{gather_}}.

\code{to_alluvia} takes a data frame with axis and axis value variables to be
used in an alluvial diagram, and reshape the data frame so that the axes
constitute separate variables whose values are given by the value variable.
This function invokes \code{\link[tidyr]{spread_}}.
}
\examples{
# Titanic data in alluvia format
titanic_alluvia <- as.data.frame(Titanic)
head(titanic_alluvia)
is_alluvial(titanic_alluvia,
            weight = "Freq")
# Titanic data in lodes format
titanic_lodes <- to_lodes(titanic_alluvia,
                          key = "x", value = "stratum", id = "alluvium",
                          axes = 1:4)
head(titanic_lodes)
is_alluvial(titanic_lodes,
            key = "x", value = "stratum", id = "alluvium",
            weight = "Freq")
# again in lodes format, this time keeping 'Class' as a variable
titanic_lodes2 <- to_lodes(titanic_alluvia,
                           key = "variable", value = "value", id = "passenger",
                           axes = 1:3, keep = "Class")
head(titanic_lodes2)
is_alluvial(titanic_lodes2,
            key = "variable", value = "value", id = "passenger",
            weight = "Freq")

# curriculum data in lodes format
data(majors)
head(majors)
is_alluvial(majors,
            key = "semester", value = "curriculum", id = "student",
            logical = FALSE)
# curriculum data in alluvia format
majors2 <- to_alluvia(majors,
                      key = "semester", value = "curriculum", id = "student")
head(majors2)
is_alluvial(majors2,
            axes = 2:9,
            logical = FALSE)

# options to relevel strata
gg <- ggplot(titanic_alluvia,
             aes(axis1 = Class, axis2 = Sex, axis3 = Age, weight = Freq))
gg +
  geom_flow(aes(fill = Survived), color = "darkgray") +
  geom_stratum() +
  geom_text(stat = "stratum", label.strata = TRUE)
gg +
  geom_flow(aes(fill = Survived), color = "darkgray", relevel.strata = TRUE) +
  geom_stratum(relevel.strata = TRUE) +
  geom_text(stat = "stratum", relevel.strata = TRUE, label.strata = TRUE)
relevs <- c("Adult", "Child")
gg +
  geom_flow(aes(fill = Survived), color = "darkgray", relevel.strata = relevs) +
  geom_stratum(relevel.strata = relevs) +
  geom_text(stat = "stratum", relevel.strata = relevs, label.strata = TRUE)
# warning when inappropriate
ggplot(titanic_lodes,
       aes(x = x, stratum = stratum, alluvium = alluvium, weight = Freq,
           label = stratum)) +
  geom_flow(color = "darkgray", relevel.strata = TRUE) +
  geom_stratum(relevel.strata = TRUE) +
  geom_text(stat = "stratum", relevel.strata = TRUE)
}
