% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvium.r
\docType{data}
\name{alluvium}
\alias{alluvium}
\alias{StatAlluvium}
\alias{stat_alluvium}
\alias{GeomAlluvium}
\alias{geom_alluvium}
\title{Alluvial flows}
\format{An object of class \code{StatAlluvium} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\arguments{
\item{width}{The width of each variable axis, as a proportion of the
separation between axes. Defaults to 1/3.}

\item{axis_width}{Deprecated; alias for \code{width}.}

\item{knot.pos}{The horizontal distance between a variable axis 
(\code{width/2} from its center) and the control point of the x-spline, as 
a proportion of the separation between the strata. (Must be between 0 and 
0.5.). Defaults to 1/6.}

\item{ribbon_bend}{Deprecated; alias for \code{knot.pos}.}

\item{lode.guidance}{The function to prioritize the axis variables for 
ordering the lodes within each stratum. Defaults to "zigzag", other options
include "rightleft", "leftright", "rightward", and "leftward" (see 
\code{\link{lode-guidance-functions}}).}

\item{bind.by.aes}{Whether to prioritize aesthetics before axes (other than 
the index axis) when ordering the lodes within each stratum. Defaults to 
FALSE.}

\item{lode.ordering}{A list (of length the number of axes) of integer vectors
(each of length the number of rows of \code{data}) or NULL entries 
(indicating no imposed ordering), or else a numeric matrix of corresponding
dimensions, giving the preferred ordering of alluvia at each axis. This 
will be used to order the lodes within each stratum by sorting the lodes 
first by stratum and then by the provided vectors.}
}
\description{
\code{stat_alluvium} calculates the depth of each group at each axis. 
\code{geom_alluvium} plots an x-spline for each group through the axes at 
these depths.
}
\section{Aesthetics}{

\code{stat_alluvium} understands only the \code{group} 
  aesthetic, but it is currently ignored.
\code{geom_alluvium} understands the following aesthetics
  (required aesthetics are in bold):
\itemize{
  \item \strong{\code{axis[0-9]*}} (\code{axis1}, \code{axis2}, etc.)
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{group}
  \item \code{linetype}
  \item \code{size}
}
Currently, \code{group} is ignored.
}

\examples{
# degeneracy (one axis; unavailable through shortcut function)
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis = Class)) +
  geom_alluvium(aes(fill = Class, alpha = Survived)) +
  scale_x_continuous(breaks = 1, labels = c("Class")) +
  scale_alpha_manual(values = c(.25, .75))

# declaration of groups (ignored)
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age,
           group = Survived)) +
  geom_alluvium() +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))

# control of horizontal spacing: parallel sets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_alluvium(aes(fill = Survived),
                width = 0, knot.pos = 0) +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))

# use of strata and labels
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_alluvium() +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age")) +
  geom_stratum() + geom_text(stat = "stratum") +
  ggtitle("Alluvial diagram of Titanic passenger demographic data")

# use of facets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex)) +
  geom_alluvium(aes(fill = Age)) +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:2, labels = c("Class", "Sex")) +
  facet_wrap(~ Survived, scales = "fixed")

# use of lode controls
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_alluvium(aes(fill = Survived),
                bind.by.aes = TRUE, lode.guidance = "rightward") +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))

# use of lode ordering
lode_ord <- replicate(n = 3, expr = sample(x = 32), simplify = FALSE)
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_alluvium(aes(fill = Survived),
                lode.ordering = lode_ord) +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))
}
\seealso{
\code{\link{stratum}} for intra-axis boxes, 
  \code{\link{alluvium_ts}} for a time series implementation, and 
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
