% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvium-ts.r
\docType{data}
\name{alluvium_ts}
\alias{GeomAlluviumTs}
\alias{StatAlluviumTs}
\alias{alluvium_ts}
\alias{geom_alluvium_ts}
\alias{stat_alluvium_ts}
\title{Alluvial flows for time series}
\format{An object of class \code{StatAlluviumTs} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\arguments{
\item{decreasing}{Logical; whether to stack the depths at each x-value with
the largest on top (FALSE, default), with the largest on bottom (TRUE), or
in the order of the grouping variable values (NA)}

\item{ribbon_bend}{The horizontal distance between a measurement time and the
control point of the x-spline, as a proportion of the separation between
times}
}
\description{
\code{stat_alluvium_ts} calculates the depth of each category at each time. 
\code{geom_alluvium_ts} plots an x-spline for each category across these 
times.
}
\examples{
# time series of WorldPhones data
wph <- as.data.frame(as.table(WorldPhones))
names(wph) <- c("Year", "Region", "Telephones")
ggplot(wph,
       aes(x = Year, group = Region, weight = Telephones)) +
  geom_alluvium_ts(aes(fill = Region, colour = Region))

# basic time series alluvia
ggplot(data = alluvial::Refugees,
       aes(x = year, weight = refugees, group = country)) +
  geom_alluvium_ts()

# time series alluvia with some aesthetics and without sorting by volume
ggplot(data = alluvial::Refugees,
       aes(x = year, weight = refugees, group = country)) +
  geom_alluvium_ts(aes(fill = country), colour = "black",
                   decreasing = NA)

# load refugees data from alluvial
data(Refugees, package = "alluvial")
# time series alluvia faceted by region
country_regions <- c(
  Afghanistan = "Middle East",
  Burundi = "Central Africa",
  `Congo DRC` = "Central Africa",
  Iraq = "Middle East",
  Myanmar = "Southeast Asia",
  Palestine = "Middle East",
  Somalia = "Horn of Africa",
  Sudan = "Central Africa",
  Syria = "Middle East",
  Vietnam = "Southeast Asia"
)
Refugees$region <- country_regions[Refugees$country]
ggplot(data = Refugees,
       aes(x = year, weight = refugees, group = country)) +
  geom_alluvium_ts(aes(fill = country), colour = "black") +
  facet_wrap(~ region, scales = "fixed")
}
\seealso{
\code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}

