% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-alluvium.r
\docType{data}
\name{geom-alluvium}
\alias{geom-alluvium}
\alias{geom_alluvium}
\alias{GeomAlluvium}
\title{Alluvial flows}
\format{An object of class \code{GeomAlluvium} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\arguments{
\item{width}{The width of each variable axis, as a proportion of the
separation between axes. Defaults to 1/3.}

\item{axis_width}{Deprecated; alias for \code{width}.}

\item{knot.pos}{The horizontal distance between a variable axis 
(\code{width/2} from its center) and the control point of the x-spline, as 
a proportion of the separation between the strata. (Must be between 0 and 
0.5.). Defaults to 1/6.}

\item{ribbon_bend}{Deprecated; alias for \code{knot.pos}.}
}
\description{
\code{stat_alluvium} calculates the depth of each group at each axis. 
\code{geom_alluvium} plots an x-spline for each group through the axes at 
these depths.
}
\section{Aesthetics}{

\code{geom_alluvium} understands the following aesthetics
  (required aesthetics are in bold):
\itemize{
  \item \strong{\code{axis[0-9]*}} (\code{axis1}, \code{axis2}, etc.)
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{group}
  \item \code{linetype}
  \item \code{size}
}
Currently, \code{group} is ignored.
}

\seealso{
\code{\link{stat_stratum}} and \code{\link{geom_stratum}} for
  intra-axis boxes, 
  \code{\link{alluvium_ts}} for a time series implementation, and 
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
