% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-alluvium.r
\docType{data}
\name{geom-alluvium}
\alias{geom-alluvium}
\alias{geom_alluvium}
\alias{GeomAlluvium}
\title{Alluvia}
\format{An object of class \code{GeomAlluvium} (inherits from \code{Geom}, \code{ggproto}) of length 7.}
\arguments{
\item{stat}{The statistical transformation to use on the data;
override the default.}

\item{width}{Numeric; the width of each stratum, as a proportion of the
distance between axes. Defaults to 1/3.}

\item{axis_width}{Deprecated; alias for \code{width}.}

\item{knot.pos}{The horizontal distance between a stratum (\code{width/2}
from its axis) and the knot of the x-spline, as a proportion of the
separation between strata. Defaults to 1/6.}

\item{ribbon_bend}{Deprecated; alias for \code{knot.pos}.}
}
\description{
\code{geom_alluvium} receives a dataset of the horizontal (\code{x}) and 
vertical (\code{y}, \code{ymin}, \code{ymax}) positions of the \strong{lodes}
of an alluvial diagram, the intersections of the alluvia with the strata.
It plots both the lodes themselves, using \code{\link{geom_lode}}, and the
flows between them, using \code{\link{geom_flow}}.
}
\section{Aesthetics}{

\code{geom_alluvium} understands the following aesthetics
  (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{y}}
  \item \strong{\code{ymin}}
  \item \strong{\code{ymax}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{linetype}
  \item \code{size}
  \item \code{group}
}
Currently, \code{group} is ignored.
}

\examples{
# basic
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age,
           fill = Survived)) +
  geom_alluvium() +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))

# time series bump chart
ggplot(alluvial::Refugees,
       aes(weight = refugees,
           x = year, stratum = country,
           fill = country, colour = country)) +
  geom_alluvium(width = 1/4, alpha = 2/3, decreasing = FALSE)

# time series line plot of refugees data, sorted by country
ggplot(data = alluvial::Refugees,
       aes(x = year, weight = refugees, stratum = country)) +
  geom_alluvium(aes(fill = country),
                colour = "black", decreasing = NA, width = 0, knot.pos = 0)
}
\seealso{
\code{\link[ggplot2]{layer}} for additional arguments,
  \code{\link{stat_stratum}} and \code{\link{geom_stratum}} for 
  intra-axis boxes, and
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
