% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-flow.r
\docType{data}
\name{geom-flow}
\alias{geom-flow}
\alias{geom_flow}
\alias{GeomFlow}
\title{Alluvial flows between lodes}
\format{An object of class \code{GeomFlow} (inherits from \code{Geom}, \code{ggproto}) of length 7.}
\arguments{
\item{stat}{The statistical transformation to use on the data;
override the default.}

\item{width}{Numeric; the width of each stratum, as a proportion of the
distance between axes. Defaults to 1/3.}

\item{axis_width}{Deprecated; alias for \code{width}.}

\item{knot.pos}{The horizontal distance between a stratum (\code{width/2}
from its axis) and the knot of the x-spline, as a proportion of the
separation between strata. Defaults to 1/6.}

\item{ribbon_bend}{Deprecated; alias for \code{knot.pos}.}

\item{aes.flow}{Character; how inter-lode flows assume aesthetics from lodes.
Options are "forward" and "backward".}
}
\description{
\code{geom_flow} receives a dataset of the horizontal (\code{x}) and 
vertical (\code{y}, \code{ymin}, \code{ymax}) positions of the \strong{lodes}
of an alluvial diagram, the intersections of the alluvia with the strata.
It reconfigures these into alluvial segments connecting pairs of
corresponding lodes in adjacent strata and plots filled x-splines between
each such pair, using a provided knot position parameter \code{knot.pos}, and
filled rectangles at either end, using a provided \code{width}.
}
\section{Aesthetics}{

\code{geom_flow} understands the following aesthetics
  (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{y}}
  \item \strong{\code{ymin}}
  \item \strong{\code{ymax}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{linetype}
  \item \code{size}
  \item \code{group}
}
Currently, \code{group} is ignored.
}

\examples{
# use of strata and labels
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_flow() +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age")) +
  geom_stratum() + geom_text(stat = "stratum") +
  ggtitle("Alluvial diagram of Titanic passenger demographic data")

# use of facets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex)) +
  geom_flow(aes(fill = Age)) +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:2, labels = c("Class", "Sex")) +
  facet_wrap(~ Survived, scales = "fixed")

# time series alluvia of WorldPhones data
wph <- as.data.frame(as.table(WorldPhones))
names(wph) <- c("Year", "Region", "Telephones")
ggplot(wph,
       aes(x = Year, stratum = Region, weight = Telephones)) +
  geom_flow(aes(fill = Region, colour = Region), width = 0)
}
\seealso{
\code{\link[ggplot2]{layer}} for additional arguments,
  \code{\link{stat_stratum}} and \code{\link{geom_stratum}} for 
  intra-axis boxes, and
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
