% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-stratum.r
\docType{data}
\name{geom-stratum}
\alias{geom-stratum}
\alias{geom_stratum}
\alias{GeomStratum}
\title{Variable axes with strata of values}
\format{An object of class \code{GeomStratum} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}) of length 6.}
\arguments{
\item{width}{The width of each stratum, as a proportion of the separation
between their centers. Defaults to 1/3.}

\item{axis_width}{Deprecated; alias for \code{width}.}
}
\description{
Given a dataset with alluvial structure, \code{geom_stratum} calculates the
centroids of the strata for each axis, together with their weights (heights)
and widths.
}
\section{Aesthetics}{

\code{geom_stratum} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
  \item \code{x}
  \item \code{y}
  \item \code{group}
  \item \code{axis[0-9]*} (\code{axis1}, \code{axis2}, etc.)
  \item \code{weight}
}
Use \code{x}, \code{y}, and \code{group} for data in lode form and 
\code{axis[0-9]*} for data in alluvium form (see \code{\link{is_alluvial}});
arguments to parameters inconsistent with the data form will be ignored.
}

\examples{
# illustrate positioning
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age, axis4 = Survived)) +
  geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:4,
                     labels = c("Class", "Sex", "Age", "Survived"))

# full axis width
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age, axis4 = Survived)) +
  geom_stratum(width = 1) + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:4,
                     labels = c("Class", "Sex", "Age", "Survived"))
  
# use of facets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex,
           fill = Survived)) +
  geom_alluvium() +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:2, labels = c("Class", "Sex")) +
  facet_wrap(~ Age, scales = "free_y")
}
\seealso{
\code{\link{stat-stratum}} for the corresponding geom.
}
\keyword{datasets}
