% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-stratum.r
\docType{data}
\name{geom-stratum}
\alias{geom-stratum}
\alias{geom_stratum}
\alias{GeomStratum}
\title{Stratum rectangles}
\format{An object of class \code{GeomStratum} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}) of length 6.}
\arguments{
\item{width}{The width of each stratum, as a proportion of the distace
between axes. Defaults to 1/3.}

\item{axis_width}{Deprecated; alias for \code{width}.}
}
\description{
\code{geom_stratum} receives a dataset of the horizontal (\code{x}) and
vertical (\code{y}, \code{ymin}, \code{ymax}) positions of the strata of an
alluvial diagram.
It plots rectangles for these strata of a provided \code{width}.
}
\section{Aesthetics}{

\code{geom_stratum} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{y}}
  \item \strong{\code{ymin}}
  \item \strong{\code{ymax}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{linetype}
  \item \code{size}
  \item \code{group}
}
Currently, \code{group} is ignored.
}

\examples{
# illustrate positioning
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age, axis4 = Survived)) +
  geom_text(stat = "stratum") +
  geom_errorbar(stat = "stratum") +
  scale_x_continuous(breaks = 1:4,
                     labels = c("Class", "Sex", "Age", "Survived"))

# full axis width
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age, axis4 = Survived)) +
  geom_stratum(width = 1) + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:4,
                     labels = c("Class", "Sex", "Age", "Survived"))
  
# use of facets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex,
           fill = Survived)) +
  geom_alluvium() +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:2, labels = c("Class", "Sex")) +
  facet_wrap(~ Age, scales = "free_y")
}
\seealso{
\code{\link{stat-stratum}} for the corresponding geom.
}
\keyword{datasets}
