% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-alluvium.r
\docType{data}
\name{geom_alluvium}
\alias{GeomAlluvium}
\alias{StatAlluvium}
\alias{StatAlluvium,}
\alias{geom_alluvium}
\alias{stat_alluvium,}
\title{Alluvial flows}
\format{An object of class \code{StatAlluvium} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\arguments{
\item{axis_width}{The width of each variable axis, as a proportion of the
separation between axes.}

\item{ribbon_bend}{The horizontal distance between a variable axis
(\code{axis_width/2} from its center) and the control point of the
x-spline, also as a proportion of the separation between the axes.}
}
\description{
\code{stat_alluvium} calculates the depth of each group at each axis. 
\code{geom_alluvium} plots an x-spline for each group through the axes at 
these depths.
}
\seealso{
\code{\link{geom_stratum}} for intra-axis boxes.
}
\keyword{datasets}

