% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-alluvium.r
\name{geom_alluvium}
\alias{geom_alluvium}
\title{Alluvia across strata}
\usage{
geom_alluvium(
  mapping = NULL,
  data = NULL,
  stat = "alluvium",
  position = "identity",
  width = 1/3,
  knot.pos = 1/4,
  knot.fix = FALSE,
  curve = "xspline",
  reach = NULL,
  segments = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data;
override the default.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{width}{Numeric; the width of each stratum, as a proportion of the
distance between axes. Defaults to 1/3.}

\item{knot.pos}{The horizontal distance between a stratum (\code{width/2}
from its axis) and the knot of the x-spline, as a proportion of the
separation between strata. Defaults to 1/6.}

\item{knot.fix}{Logical; whether to interpret \code{knot.pos} as a fixed value,
rather than (the default) as a proportion of the separation between
adjacent axes.}

\item{curve}{Character; the type of curve used to produce flows. Defaults to
\code{"xspline"} and can be alternatively set to one of \code{"linear"}, \code{"cubic"},
\code{"quintic"}, \code{"sine"}, \code{"arctangent"}, and \code{"sigmoid"}. \code{"xspline"}
produces approximation splines using 4 points per curve; the alternatives
produce interpolation splines between points along the graphs of functions
of the associated type. See the \strong{Curves} section.}

\item{reach}{For alternative \code{curve}s based on asymptotic functions, the
value along the asymptote at which to truncate the function to obtain the
shape that will be scaled to fit between strata. See the \strong{Curves}
section.}

\item{segments}{The number of segments to be used in drawing each alternative
curve (each curved boundary of each flow). If less than 3, will be silently
changed to 3.}

\item{na.rm}{Logical:
if \code{FALSE}, the default, \code{NA} lodes are not included;
if \code{TRUE}, \code{NA} lodes constitute a separate category,
plotted in grey (regardless of the color scheme).}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\description{
\code{geom_alluvium} receives a dataset of the horizontal (\code{x}) and
vertical (\code{y}, \code{ymin}, \code{ymax}) positions of the \strong{lodes}
of an alluvial plot, the intersections of the alluvia with the strata.
It plots both the lodes themselves, using \code{\link[=geom_lode]{geom_lode()}}, and the
flows between them, using \code{\link[=geom_flow]{geom_flow()}}.
}
\section{Aesthetics}{

\code{geom_alluvium}, \code{geom_flow}, \code{geom_lode}, and \code{geom_stratum} understand the
following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{\code{ymin}}
\item \strong{\code{ymax}}
\item \code{alpha}
\item \code{colour}
\item \code{fill}
\item \code{linetype}
\item \code{size}
\item \code{group}
}

\code{group} is used internally; arguments are ignored.
}

\section{Curves}{

By default, \code{geom_alluvium()} and \code{geom_flow()} render flows between lodes as
splines about quartets of control points using \code{\link[grid:xsplineGrob]{grid::xsplineGrob()}}. These
curves are parameterized by the relative location of the knot (\code{knot.pos})
that governs the bending of each end toward their meeting at the point of
inflection. They are quick to render and clear to read. Nevertheless, for any
number of reasons, users may prefer plots that use differently-shaped curves.

A variety of such options are illustrated in \href{https://www.dataplusscience.com/Sigmoid.html}{this blog post by Jeffrey Shaffer}. A subset of these,
some with added flexibility, are implemented here. Each is encoded as a
continuous, increasing, bijective function from the unit interval \eqn{[0,1]}
to itself, and each is rescaled, if necessary, so that its endpoints meet the
corresponding lodes. They are rendered piecewise-linearly, by default using
\code{segments = 48}. Summon each curve type by passing one of the following
strings to \code{curve}:
\itemize{
\item \code{"linear"}:     \eqn{f(x)=x}, the unique degree-1 polynomial that takes
0 to 0 and 1 to 1
\item \code{"cubic"}:      \eqn{f(x)=3x^{2}-2x^{3}}{f(x)=3x^2-2x^3}, the unique
degree-3 polynomial that also is flat at both endpoints
\item \code{"quintic"}:    \eqn{f(x)=10x^{3}-15x^{4}+6x^{5}}{f(x)=10x^3-15x^4+6x^5},
the unique degree-5 polynomial that also is flat at both
endpoints
\item \code{"sine"}:       the unique sinusoidal function that is flat at both
endpoints
\item \code{"arctangent"}: the inverse tangent function, scaled and re-centered to the
unit interval from a symmetric domain with radius \code{reach}
\item \code{"sigmoid"}:    the sigmoid function, scaled and re-centered to the unit
interval from a symmetric domain with radius \code{reach}
}

Only the (default) \code{"xspline"} option uses the \verb{knot.*} parameters, while
only the alternative curves use the \code{segments} parameter, and only
\code{"arctangent"} and \code{"sigmoid"} use the \code{reach} parameter. Larger values of
\code{reach} result in greater compression and steeper slopes. The \code{NULL} default
will be changed to \code{2+sqrt(3)} for \code{"arctangent"} and to \code{6} for \code{"sigmoid"}.
}

\section{Defunct parameters}{

The previously defunct parameters \code{axis_width} and \code{ribbon_bend} have been
discontinued. Use \code{width} and \code{knot.pos} instead.
}

\examples{
# basic
ggplot(as.data.frame(Titanic),
       aes(y = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age,
           fill = Survived)) +
  geom_alluvium() +
  scale_x_discrete(limits = c("Class", "Sex", "Age"))

gg <- ggplot(alluvial::Refugees,
             aes(y = refugees, x = year, alluvium = country))
# time series bump chart (quintic flows)
gg + geom_alluvium(aes(fill = country, colour = country),
                   width = 1/4, alpha = 2/3, decreasing = FALSE,
                   curve = "quintic")
# time series line plot of refugees data, sorted by country
gg + geom_alluvium(aes(fill = country, colour = country),
                   decreasing = NA, width = 0, knot.pos = 0)

# irregular spacing between axes of a continuous variable
data(Refugees, package = "alluvial")
refugees_sub <- subset(Refugees, year \%in\% c(2003, 2005, 2010, 2013))
gg <- ggplot(data = refugees_sub,
             aes(x = year, y = refugees, alluvium = country)) +
  theme_bw() +
  scale_fill_brewer(type = "qual", palette = "Set3")
# proportional knot positioning (default)
gg +
  geom_alluvium(aes(fill = country),
                alpha = .75, decreasing = FALSE) +
  geom_stratum(aes(stratum = country), decreasing = FALSE, width = 1/2)
# fixed knot positioning
gg +
  geom_alluvium(aes(fill = country),
                alpha = .75, decreasing = FALSE,
                knot.pos = 1, knot.fix = TRUE) +
  geom_stratum(aes(stratum = country), decreasing = FALSE, width = 1/2)
# coarsely-segmented curves
gg +
  geom_alluvium(aes(fill = country),
                alpha = .75, decreasing = FALSE,
                curve = "arctan", segments = 6) +
  geom_stratum(aes(stratum = country), decreasing = FALSE, width = 1/2)
}
\seealso{
\code{\link[ggplot2:layer]{ggplot2::layer()}} for additional arguments and
\code{\link[=stat_alluvium]{stat_alluvium()}} and
\code{\link[=stat_flow]{stat_flow()}} for the corresponding stats.

Other alluvial geom layers: 
\code{\link{geom_flow}()},
\code{\link{geom_lode}()},
\code{\link{geom_stratum}()}
}
\concept{alluvial geom layers}
