% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-axes.r
\name{get_axes}
\alias{get_axes}
\title{Identify axis columns}
\usage{
get_axes(x)
}
\arguments{
\item{x}{Character vector}
}
\description{
Identify elements in a character vector that fit the pattern of axis
aesthetic names, and return their indices in the numerical order of the axis
numbers.
}

