% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel.r
\name{ggalluvial-deprecated}
\alias{ggalluvial-deprecated}
\title{Deprecated functions}
\description{
These functions are deprecated in the current version and will be removed for
the eventual CRAN release.
}
\keyword{internal}
