% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalluvial-package.r
\docType{package}
\name{ggalluvial-package}
\alias{ggalluvial}
\alias{ggalluvial-package}
\title{ggalluvial: Alluvial Plots in 'ggplot2'}
\description{
Alluvial plots use x-splines, sometimes augmented with stacked
    histograms, to visualize multi-dimensional or repeated-measures data with
    categorical or ordinal variables. They can be viewed as simplified and
    standardized Sankey diagrams; see Riehmann, Hanfler, and Froehlich (2005)
    <doi:10.1109/INFVIS.2005.1532152> and Rosvall and Bergstrom (2010)
    <doi:10.1371/journal.pone.0008694>. This package provides ggplot2 layers
    to produce alluvial plots from tidy data.
}
\section{Acknowledgments}{


Many users identified problems and suggested improvements via email and the
GitHub issue tracker.

Development benefitted from the use of equipment and the support of
colleagues at \href{https://health.uconn.edu/}{UConn Health}.
}

\seealso{
Useful links:
\itemize{
  \item \url{http://corybrunson.github.io/ggalluvial/}
  \item Report bugs at \url{https://github.com/corybrunson/ggalluvial/issues}
}

}
\author{
\strong{Maintainer}: Jason Cory Brunson \email{cornelioid@gmail.com}

}
\keyword{internal}
