% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalluvial.r
\name{ggalluvial}
\alias{ggalluvial}
\alias{ggalluvial_lodes}
\alias{ggalluvial_alluvia}
\alias{ggalluvial_formula}
\title{Formula interface for alluvial diagrams}
\usage{
ggalluvial(...)

ggalluvial_lodes(..., incl.strata = TRUE, incl.labels = FALSE)

ggalluvial_alluvia(..., incl.strata = TRUE, incl.labels = FALSE)

ggalluvial_formula(formula, data = NULL, weight, incl.strata = TRUE,
  incl.labels = TRUE, ...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot} and inherited by 
\code{geom_alluvium} and \code{geom_stratum}.}

\item{incl.strata}{Logical; whether to plot strata over alluvia.}

\item{incl.labels}{Logical; whether to plot labels over strata.}

\item{formula}{A formula to specify the axes and alluvial divisions.}

\item{data}{A data frame or frequency table.}

\item{weight}{A weight variable, from \code{data} or of compatible length 
with the elements of \code{formula}.}
}
\description{
\code{ggalluvial} produces an alluvial diagram with flows, boxes, and labels,
optionally based on a formula in terms of the data elements.
}
\seealso{
\code{\link{stat_stratum}}, \code{\link{geom_stratum}},
  \code{\link{stat_alluvium}}, and \code{\link{geom_alluvium}}
}
