% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{majors}
\alias{majors}
\title{Students' declared majors across several semesters}
\format{
A data frame with 80 rows and 3 variables:
\describe{
\item{\code{student}}{student identifier}
\item{\code{semester}}{character tag for odd-numbered semesters}
\item{\code{curriculum}}{declared major program}
}
}
\description{
This data set follows the major curricula of 10 students across 8 academic
semesters. Missing values indicate undeclared majors. The data were kindly
contributed by Dario Bonaretti.
}
\keyword{datasets}
