% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize-overlaps.r
\name{minimize-overlaps}
\alias{minimize-overlaps}
\alias{optimize_strata}
\alias{objective_fun}
\alias{permute_strata}
\alias{permute_axis_strata}
\title{Optimize strata}
\usage{
optimize_strata(data, key, value, id, weight = NULL, objective = "count",
  method = NULL, niter = 6)

objective_fun(data, key, value, id, weight, perm)

permute_strata(data, key, value, id, perm)

permute_axis_strata(data, key, value, id, perms)
}
\arguments{
\item{data}{Data frame in lode form (see \code{\link{to_lodes}}).}

\item{key, value, id}{Numeric or character; the fields of \code{data}
corresponding to the axis (variable), stratum (value), and alluvium
(identifying) variables.}

\item{weight}{Optional numeric or character; the fields of \code{data}
corresponding to alluvium or lode weights (heights when plotted). The
default value \code{NULL} assigns every alluvium equal weight.}

\item{objective}{Character, the objective function to minimize; matched to
\code{"count"} or \code{"weight"}.}

\item{method}{Character; whether to exhaust all permutations of all axes
(\code{"exhaustive"}) versus to perform heuristic optimization from random
initial permutations (\code{"heuristic"}). Default, \code{NULL}, is to use
the heuristic algorithm unless no axis has six or more strata.}

\item{niter}{Positive integer; if \code{method} is \code{"heuristic"}, the
number of iterations to perform from randomly sampled seed permutation
sets.}

\item{perm, perms}{An integer vector or list of integer vectors encoding 
permutations of the strata, either all together (as output by 
\code{optimize_strata()}) or for each axis (as output by
\code{optimize_axis_strata()}).}
}
\description{
The function \code{optimize_strata()} attempts to order the strata at each 
axis so as to maximize the readability of the diagram, by minimizing an 
objective function related to the overlaps of crossing flows. The objective
functions include the following:
}
\details{
\itemize{
  \item \code{"count"}: the number of crossings
  \item \code{"weight"}: the total weight of the crossings
  (the weight of each crossing is the product of the weights of the flows)
}

The function \code{permute_strata()} then applies the stratum permutations to
a data frame (which must have the correct number of strata at each axis).
}
