% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize-overlaps.r
\name{minimize-overlaps}
\alias{minimize-overlaps}
\alias{optimize_strata}
\alias{permute_strata}
\title{Optimize strata}
\usage{
optimize_strata(data, key, value, id, weight = NULL, free.strata = TRUE,
  objective = "count", method = "exhaustive", niter = 6)

permute_strata(data, key, value, id, permutations)
}
\arguments{
\item{data}{Data frame in lode form (see \code{\link{to_lodes}}).}

\item{key, value, id}{Numeric or character; the fields of \code{data}
corresponding to the axis (variable), stratum (value), and alluvium
(identifying) variables.}

\item{weight}{Optional numeric or character; the fields of \code{data}
corresponding to alluvium or lode weights (heights when plotted). The
default value \code{NULL} assigns every alluvium equal weight.}

\item{free.strata}{Logical; whether to allow the same strata to be ordered
differently at different axes.}

\item{objective}{Character, the objective function to minimize; matched to
\code{"count"} or \code{"weight"}.}

\item{method}{Character; whether to exhaust all permutations of all axes
(\code{"exhaustive"}) versus to perform heuristic optimization from random
initial permutations (\code{"heuristic"}).}

\item{niter}{Positive integer; if \code{method == "heuristic"}, the number of
iterations to perform from randomly sampled seed permutation sets.}

\item{permutations}{A list of integer vectors, one for each axis, each a
permutation of the number of strata at the corresponding axis (as output by
\code{optimize_strata()}).}
}
\description{
The function \code{optimize_strata()} attempts to order the strata at each 
axis so as to maximize the readability of the diagram, by minimizing an 
objective function related to the overlaps of crossing flows. The objective
functions include the following:
}
\details{
\itemize{
  \item \code{"count"}: the number of crossings
  \item \code{"weight"}: the total weight of the crossings
  (the weight of each crossing is the product of the weights of the flows)
}

The function \code{permute_strata()} then applies the stratum permutations to
a data frame (which must have the correct number of strata at each axis).
}
