% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-alluvium.r
\docType{data}
\name{stat-alluvium}
\alias{stat-alluvium}
\alias{stat_alluvium}
\alias{StatAlluvium}
\title{Alluvial flows}
\format{An object of class \code{StatAlluvium} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\arguments{
\item{lode.guidance}{The function to prioritize the axis variables for 
ordering the lodes within each stratum. Defaults to "zigzag", other options
include "rightleft", "leftright", "rightward", and "leftward" (see 
\code{\link{lode-guidance-functions}}).}

\item{bind.by.aes}{Whether to prioritize aesthetics before axes (other than 
the index axis) when ordering the lodes within each stratum. Defaults to 
FALSE.}

\item{lode.ordering}{A list (of length the number of axes) of integer vectors
(each of length the number of rows of \code{data}) or NULL entries 
(indicating no imposed ordering), or else a numeric matrix of corresponding
dimensions, giving the preferred ordering of alluvia at each axis. This 
will be used to order the lodes within each stratum by sorting the lodes 
first by stratum and then by the provided vectors.}
}
\description{
\code{stat_alluvium} calculates the depth of each group at each axis. 
\code{geom_alluvium} plots an x-spline for each group through the axes at 
these depths.
}
\section{Aesthetics}{

\code{stat_alluvium} understands only the \code{group} 
  aesthetic, but it is currently ignored.
}

\seealso{
\code{\link{stat_stratum}} and \code{\link{geom_stratum}} for
  intra-axis boxes, 
  \code{\link{alluvium_ts}} for a time series implementation, and 
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
