% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-alluvium.r
\docType{data}
\name{stat-alluvium}
\alias{stat-alluvium}
\alias{stat_alluvium}
\alias{StatAlluvium}
\title{Alluvial positions}
\format{An object of class \code{StatAlluvium} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\arguments{
\item{lode.guidance}{The function to prioritize the axis variables for 
ordering the lodes within each stratum. Defaults to "zigzag", other options
include "rightleft", "leftright", "rightward", and "leftward" (see 
\code{\link{lode-guidance-functions}}).}

\item{aes.bind}{Whether to prioritize aesthetics before axes (other than the
index axis) when ordering the lodes within each stratum. Defaults to FALSE.}

\item{lode.ordering}{A list (of length the number of axes) of integer vectors
(each of length the number of rows of \code{data}) or NULL entries 
(indicating no imposed ordering), or else a numeric matrix of corresponding
dimensions, giving the preferred ordering of alluvia at each axis. This 
will be used to order the lodes within each stratum by sorting the lodes 
first by stratum and then by the provided vectors.}
}
\description{
Given a dataset with alluvial structure, \code{stat_alluvium} calculates the 
centroids (\code{x} and \code{y}) of the \strong{lodes}, the intersections of
the alluvia with the strata, together with their weights (heights; 
\code{ymin} and \code{ymax}). It leverages the \code{group} aesthetic for 
plotting purposes (for now).
}
\section{Aesthetics}{
 \code{stat_alluvium} understands the following
  aesthetics (required aesthetics are in bold):
\itemize{
  \item \code{x}
  \item \code{stratum}
  \item \code{alluvium}
  \item \code{axis[0-9]*} (\code{axis1}, \code{axis2}, etc.)
  \item \code{weight}
  \item \code{group}
}
Currently, \code{group} is ignored.
Use \code{x}, \code{stratum}, and \code{alluvium} for data in lode form and 
\code{axis[0-9]*} for data in alluvium form (see \code{\link{is_alluvial}});
arguments to parameters inconsistent with the data format will be ignored.
}

\examples{
# illustrate positioning
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age,
           color = Survived)) +
  geom_errorbar(stat = "stratum") +
  geom_line(stat = "alluvium") +
  geom_pointrange(stat = "alluvium") +
  geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:3,
                     labels = c("Class", "Sex", "Age"))

# declaration of groups (ignored)
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age,
           group = Survived)) +
  geom_alluvium() +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))

# parallel sets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_alluvium(aes(fill = Survived),
                width = 1/8, knot.pos = 0) +
  geom_stratum(width = 1/8) + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age")) +
  coord_flip()

# use of strata and labels
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_flow() +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age")) +
  geom_stratum() + geom_text(stat = "stratum") +
  ggtitle("Alluvial diagram of Titanic passenger demographic data")

# use of facets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex)) +
  geom_flow(aes(fill = Age)) +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:2, labels = c("Class", "Sex")) +
  facet_wrap(~ Survived, scales = "fixed")

# use of lode controls
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_flow(aes(fill = Survived),
            aes.bind = TRUE, lode.guidance = "rightward") +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))

# use of lode ordering
lode_ord <- replicate(n = 3, expr = sample(x = 32), simplify = FALSE)
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_flow(aes(fill = Survived),
            lode.ordering = lode_ord) +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))
# time series bump chart
ggplot(alluvial::Refugees,
       aes(weight = refugees,
           x = year, stratum = country, alluvium = country,
           fill = country)) +
  geom_alluvium(width = 1/4, alpha = 2/3, decreasing = FALSE)

# time series alluvia of WorldPhones data
wph <- as.data.frame(as.table(WorldPhones))
names(wph) <- c("Year", "Region", "Telephones")
ggplot(wph,
       aes(x = Year, stratum = Region, alluvium = Region, weight = Telephones)) +
  geom_flow(aes(fill = Region, colour = Region), width = 0)

# time series line plot of refugees data, sorted by country
ggplot(data = alluvial::Refugees,
       aes(x = year, weight = refugees, stratum = country, alluvium = country)) +
  geom_flow(aes(fill = country),
            colour = "black", decreasing = NA, width = 0, knot.pos = 0)

# load refugees data from alluvial
data(Refugees, package = "alluvial")
# time series alluvia faceted by region
country_regions <- c(
  Afghanistan = "Middle East",
  Burundi = "Central Africa",
  `Congo DRC` = "Central Africa",
  Iraq = "Middle East",
  Myanmar = "Southeast Asia",
  Palestine = "Middle East",
  Somalia = "Horn of Africa",
  Sudan = "Central Africa",
  Syria = "Middle East",
  Vietnam = "Southeast Asia"
)
Refugees$region <- country_regions[Refugees$country]
ggplot(data = Refugees,
       aes(x = year, weight = refugees, stratum = country, alluvium = country)) +
  geom_alluvium(aes(fill = country), decreasing = FALSE) +
  facet_wrap(~ region, scales = "fixed")
}
\seealso{
\code{\link{geom_alluvium}} for the corresponding geom,
  \code{\link{stat_stratum}} and \code{\link{geom_stratum}} for
  intra-axis boxes, and
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
