% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-stratum.r
\docType{data}
\name{stat-stratum}
\alias{stat-stratum}
\alias{stat_stratum}
\alias{StatStratum}
\title{Stratum positions}
\format{An object of class \code{StatStratum} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\arguments{
\item{geom}{The geometric object to use display the data;
override the default.}

\item{decreasing}{Logical; whether to arrange the strata at each axis
in the order of the variable values (\code{NA}, the default),
in ascending order of total weight (largest on top, \code{FALSE}), or
in descending order of total weight (largest on bottom, \code{TRUE}).}

\item{reverse}{Logical; if \code{decreasing} is \code{NA},
whether to arrange the strata at each axis
in the reverse order of the variable values,
so that they match the order of the values in the legend.
Ignored if \code{decreasing} is not \code{NA}.
Defaults to \code{TRUE}.}

\item{label.strata}{Logical; whether to assign the values of the axis
variables to the strata. Defaults to FALSE, and requires that no label
aesthetic is assigned.}
}
\description{
Given a dataset with alluvial structure, \code{stat_stratum} calculates the
centroids of the strata at each axis, together with their weights (heights).
}
\section{Aesthetics}{

\code{stat_stratum} understands the following aesthetics:
\itemize{
  \item \code{x}
  \item \code{stratum}
  \item \code{alluvium}
  \item \code{axis[0-9]*} (\code{axis1}, \code{axis2}, etc.)
  \item \code{weight}
  \item \code{group}
}
Currently, \code{group} is ignored.
Use \code{x}, \code{stratum}, and (optionally) \code{alluvium} for data in
lode form and \code{axis[0-9]*} for data in alluvium form (see
\code{\link{is_alluvial}}); arguments to parameters inconsistent with the
data format will be ignored.
}

\examples{
# only 'stratum' assignment is necessary to generate strata
data(vaccinations)
ggplot(vaccinations,
       aes(weight = freq,
           x = survey, stratum = response,
           fill = response)) +
  stat_stratum(width = .5)

# lode data: positioning with weight labels
ggplot(vaccinations,
       aes(weight = freq,
           x = survey, stratum = response, alluvium = subject,
           label = freq)) +
  stat_stratum(geom = "errorbar") +
  geom_text(stat = "stratum")
# lode data: positioning with stratum labels
ggplot(vaccinations,
       aes(weight = freq,
           x = survey, stratum = response, alluvium = subject,
           label = response)) +
  stat_stratum(geom = "errorbar") +
  geom_text(stat = "stratum")

# alluvium data: positioning with weight labels
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age, axis4 = Survived,
           label = Freq)) +
  geom_text(stat = "stratum") +
  stat_stratum(geom = "errorbar") +
  scale_x_continuous(breaks = 1:4,
                     labels = c("Class", "Sex", "Age", "Survived"))
# alluvium data: positioning with stratum labels
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age, axis4 = Survived)) +
  geom_text(stat = "stratum", label.strata = TRUE) +
  stat_stratum(geom = "errorbar") +
  scale_x_continuous(breaks = 1:4,
                     labels = c("Class", "Sex", "Age", "Survived"))
}
\seealso{
\code{\link[ggplot2]{layer}} for additional arguments,
  \code{\link{geom_stratum}} for the corresponding geom,
  \code{\link{stat_alluvium}} and \code{\link{geom_alluvium}} for
  alluvial flows, and
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
