% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-alluvium.r
\docType{data}
\name{stat_alluvium}
\alias{StatAlluvium}
\alias{stat_alluvium}
\title{Alluvial flows}
\format{An object of class \code{StatAlluvium} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\arguments{
\item{axis_width}{The width of each variable axis, as a proportion of the 
separation between axes.}
}
\description{
\code{stat_alluvium} calculates the depth of each group at each axis, for the
purpose of tracing flows using \code{\link{geom_alluvium}}.
}
\section{Aesthetics}{
 \code{stat_alluvium} understands only the \code{group} 
  aesthetic, but it is currently ignored.
}
\examples{
# basic flows (alluvia)
ggplot(as.data.frame(Titanic),
       aes(weight = Freq, axis1 = Class, axis2 = Sex)) +
  geom_alluvium()

# use of facets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_alluvium(aes(fill = Age, alpha = Sex)) +
  facet_wrap(~ Survived, scales = "free_y")

# declaration of groups (ignored)
ggplot(as.data.frame(Titanic),
       aes(weight = Freq, axis1 = Class, axis2 = Sex, group = Survived)) +
  geom_alluvium()

# degeneracy (one axis; unavailable through shortcut function)
ggplot(as.data.frame(Titanic), aes(weight = Freq, axis = Class)) +
  geom_alluvium(aes(fill = Class, color = Survived)) +
  scale_color_manual(values = c("black", "white"))

# degeneracy (no axis)
if (FALSE) {
  ggplot(as.data.frame(Titanic), aes(weight = Freq)) +
    geom_alluvium(aes(fill = Class, color = Survived)) +
    scale_color_manual(values = c("black", "white"))
}

# control of horizontal spacing
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age, axis4 = Survived)) +
  geom_alluvium(aes(fill = Age),
                axis_width = 1/5, ribbon_bend = 1/3)

# control of axis positions (ridiculous syntax)
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis1.5 = Age, axis2.5 = Sex, axis3 = Survived)) +
  geom_alluvium(aes(fill = Age))

# use of annotation and labels
ggplot(as.data.frame(Titanic),
       aes(weight = Freq, axis1 = Class, axis2 = Sex, axis3 = Age)) +
  geom_alluvium() +
  geom_text(stat = "stratum") +
  ggtitle("Alluvial diagram of Titanic passenger demographic data") +
  scale_x_continuous(breaks = 1:3, labels = c("Class", "Sex", "Age"))

# combining flows and boxes
ggplot(as.data.frame(UCBAdmissions),
       aes(weight = Freq, axis1 = Gender, axis2 = Dept)) +
  geom_alluvium(aes(fill = Admit)) +
  geom_stratum() + geom_text(stat = "stratum")
}
\seealso{
\code{\link{geom_alluvium}} for plotting alluvia from transformed
  data, \code{\link{stat_stratum}} and \code{\link{geom_stratum}} for
  intra-axis boxes, and \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}

