% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-stratum.r
\docType{data}
\name{stat_stratum}
\alias{StatStratum}
\alias{stat_stratum}
\title{Variable axes and strata}
\format{An object of class \code{StatStratum} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\arguments{
\item{axis_width}{The width of each variable axis, as a proportion of the 
separation between axes.}
}
\description{
\code{stat_stratum} calculates the centers of the levels at each axis.
}
\section{Aesthetics}{
 \code{stat_stratum} understands only the \code{group}
  aesthetic, but it is currently ignored.
}
\examples{
# basic boxes (strata) with labels
ggplot(as.data.frame(Titanic),
       aes(weight = Freq, axis1 = Class, axis2 = Sex)) +
  geom_stratum() + geom_text(stat = "stratum")

# full axis width
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age, axis4 = Survived)) +
  geom_stratum(axis_width = 1) + geom_text(stat = "stratum")

# adding flows to boxes
ggplot(as.data.frame(Titanic),
       aes(weight = Freq, axis1 = Class, axis2 = Sex)) +
  geom_stratum() + geom_text(stat = "stratum") +
  geom_alluvium(aes(fill = Age))

# combining flows and boxes and using facets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq, axis1 = Class, axis2 = Sex)) +
  geom_alluvium(aes(fill = Age)) +
  geom_stratum() + geom_text(stat = "stratum") +
  facet_wrap(~ Survived, scales = "free_y")

# declaration of groups (ignored)
ggplot(as.data.frame(Titanic),
       aes(weight = Freq, axis1 = Class, axis2 = Sex, group = Survived)) +
  geom_alluvium() +
  geom_stratum() + geom_text(stat = "stratum")
}
\seealso{
\code{\link{geom_stratum}} for plotting strata from transformed
  data, \code{\link{stat_alluvium}} and \code{\link{geom_alluvium}} for
  inter-axis flows and \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}

