% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratum.r
\docType{data}
\name{stratum}
\alias{stratum}
\alias{StatStratum}
\alias{stat_stratum}
\alias{GeomStratum}
\alias{geom_stratum}
\title{Variable axes and strata}
\format{An object of class \code{StatStratum} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\arguments{
\item{axis_width}{The width of each variable axis, as a proportion of the 
separation between axes.}
}
\description{
\code{stat_stratum} calculates the centers of the levels at each axis. 
\code{geom_stratum} stacks a box for each level of a variable at its axis.
}
\section{Aesthetics}{

\code{stat_stratum} understands only the \code{group} aesthetic, but it is
currently ignored.
\code{geom_stratum} understands the following aesthetics (required
aesthetics are in bold):
\itemize{
  \item \strong{\code{xmin}}
  \item \strong{\code{xmax}}
  \item \strong{\code{ymin}}
  \item \strong{\code{ymax}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{group}
  \item \code{linetype}
  \item \code{size}
}
Currently, \code{group} is ignored.
}

\examples{
# full axis width
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex, axis3 = Age, axis4 = Survived)) +
  geom_stratum(axis_width = 1) + geom_text(stat = "stratum")

# use of facets
ggplot(as.data.frame(Titanic),
       aes(weight = Freq,
           axis1 = Class, axis2 = Sex,
           fill = Survived)) +
  geom_alluvium() +
  geom_stratum() + geom_text(stat = "stratum") +
  scale_x_continuous(breaks = 1:2, labels = c("Class", "Sex")) +
  facet_wrap(~ Age, scales = "free_y")
}
\seealso{
\code{\link{alluvium}} for inter-axis flows and
  \code{\link{ggalluvial}} for a shortcut method.
}
\keyword{datasets}
