% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{vaccinations}
\alias{vaccinations}
\title{Influenza vaccination rates}
\format{An alluvial data frame in lodes form.}
\source{
RAND American Life Panel \url{https://alpdata.rand.org/}
}
\description{
This is a sample from the RAND ALP surveys on influenza vaccination,
kindly contributed by Raffaele Vardavas.
}
